Add files via upload
I realized that the MobileNetV2 model is not suitable for face recognition, so the algorithm is correct, but the output it gives does not work for the purpose I want. The next project I will share will be made using only the opencv plugin and will be suitable for the purpose.